lsm: start bbpp
BBPP will be the same as bipbuf, but using a pool of "protectors" to allow a
pool of readers. Each reader will get a "protector" which tells the
potential writer that it can't modify the protected region.
The reader will increment the index on its protector as it scans the buffer.
By checking all protectors, the writer can see the safe maximum index to
which it can modify the buffer.